Update excludes for new Frechet tests 102 (#103)
* Update test excludes for new Frechet tests

* Fix whitespace

* Update hootenanny.spec.in

Walking through the build process (not fun) led me to this also needing to be changed.  This is where the error is happening.  I think that the other is once the RPMs are installed then it runs the tests again.  This test is creating the RPM for Hootenanny.

* Fixed bad exclude regex

Missing an "S" in the exclude regex...


Former-commit-id: 8f32a6be6ef3cf4cd8b9dc3af1e8a338926eb0ca
Former-commit-id: 2d9210c3ebd80599bbf6d034b9b46bdae2a3a391
Former-commit-id: 7cedf23e2e33b574fa71e81ed8be4d3172c0a9fe [formerly 618bc02ae3306886bcf5e5ea80ebfeb6e3465792]
Former-commit-id: 47548a5e8b171acdaeb4e45f6dcf724e293bca52
Former-commit-id: 6fa6428d7e1346760e54885f4c83588d5e79a6ec [formerly 190a50cd967fac8a10e32fada830b0352c6450cc]
Former-commit-id: e6170e61a7305a91662bdbaad1b690a76f19895f