Update excludes for new Frechet tests 102 (#103)
* Update test excludes for new Frechet tests

* Fix whitespace

* Update hootenanny.spec.in

Walking through the build process (not fun) led me to this also needing to be changed.  This is where the error is happening.  I think that the other is once the RPMs are installed then it runs the tests again.  This test is creating the RPM for Hootenanny.

* Fixed bad exclude regex

Missing an "S" in the exclude regex...


Former-commit-id: 369c1b2ec3966b7f24054cfa3716ca5781499f2e
Former-commit-id: 8daf39db3054a6af6f745a5d4a4040839b385931
Former-commit-id: 0ad43f535cf9d776ef15e20103a48c25f957d882 [formerly b6c03e7ba7e4dc7d9bb0060ce4455396fabc1ed6]
Former-commit-id: 2717ffedea11a58e249e90d21eac52203c718a4b
Former-commit-id: 97a5bc8c33b39e794b8bbac977d8a4496972df77 [formerly c11f457694a1b7828c417427409112e1cf3a87b7]
Former-commit-id: 228e1d4d3e84afc8ae69298343799aee086f9dfb