Merge latest jdk 131 fix into develop (#149)
* Update excludes for new Frechet tests 102 (#103)

* Update test excludes for new Frechet tests

* Fix whitespace

* Update hootenanny.spec.in

Walking through the build process (not fun) led me to this also needing to be changed.  This is where the error is happening.  I think that the other is once the RPMs are installed then it runs the tests again.  This test is creating the RPM for Hootenanny.

* Fixed bad exclude regex

Missing an "S" in the exclude regex...

* Fix jdk requires to latest release


Former-commit-id: 678d6fd30004a0cf4f304b26dc57c9a50ae5c5c6
Former-commit-id: 1160a5c78c35670a72c4de88b7036f30307590a8
Former-commit-id: 01b4e886b950a6c48b878a4f0367da06b409055b [formerly b408531c599664a47f3f59fdece3c94437ea1793]
Former-commit-id: 43c30be24a13144772e50ae883b8d899c52fef52
Former-commit-id: 63edb515a40f129b3649feb41561ab98eb633dfd [formerly 79ffecd4f3a57ef389e3d44a2616858122c8fb5b]
Former-commit-id: c726ae41dc58a144e2e987d0e9b8a6d5bd4e7174